Support for default outputType and trackingId
